third_party/WebKit/LayoutTests/fast/forms/text-area-resize-orthogonal-containingBlock-expected.txt

Test for resizing the Textarea below it's initial size and with orthogonal containing block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("textInputID").style.width is "64px"
PASS document.getElementById("textInputID").style.height; is "74px"
PASS successfullyParsed is true

TEST COMPLETE
